♻️ refactor: rename the agentConfig folder to agent
