+++ src/hooks/useSTT.ts
@@ -12,7 +12,7 @@ import { createHeaderWithOpenAI } from '@/services/_header';
 import { OPENAI_URLS } from '@/services/_url';
 import { settingsSelectors, useGlobalStore } from '@/store/global';
 import { useSessionStore } from '@/store/session';
-import { agentSelectors } from '@/store/session/slices/agentConfig';
+import { agentSelectors } from '@/store/session/selectors';
 
 interface STTConfig extends SWRConfiguration {
   onTextChange: (value: string) => void;
