+++ src/hooks/useTTS.ts
@@ -14,7 +14,7 @@ import { createHeaderWithOpenAI } from '@/services/_header';
 import { OPENAI_URLS, TTS_URL } from '@/services/_url';
 import { settingsSelectors, useGlobalStore } from '@/store/global';
 import { useSessionStore } from '@/store/session';
-import { agentSelectors } from '@/store/session/slices/agentConfig';
+import { agentSelectors } from '@/store/session/selectors';
 import { TTSServer } from '@/types/session';
 
 interface TTSConfig extends SWRConfiguration {
