+++ src/store/session/selectors.ts
@@ -1,3 +1,3 @@
-export { agentSelectors } from './slices/agentConfig/selectors';
+export { agentSelectors } from './slices/agent/selectors';
 export { chatSelectors, topicSelectors } from './slices/chat/selectors';
 export { sessionSelectors } from './slices/session/selectors';
