+++ src/store/session/slices/agent/selectors.ts
